implementing l2dct

+++ src/internet/model/d2tcp-socket.cc
@@ -220,7 +220,7 @@ D2tcpSocket::SlowDown (void)
         }
     }
   double p = std::pow (m_alpha, d);
-  uint32_t newCwnd =  (uint32_t)((1 - p / 2.0) * m_tcb->m_cWnd);
+  uint32_t newCwnd =  (1 - p / 2.0) * m_tcb->m_cWnd;
   // halve cwnd according to D2TCP algo
   m_tcb->m_ssThresh = std::max (newCwnd, 2 * m_tcb->m_segmentSize);
   m_tcb->m_cWnd = std::max (newCwnd, m_tcb->m_segmentSize);
